+++ gradio/components/datetime.py
@@ -10,8 +10,10 @@ import pytz
 from gradio_client.documentation import document
 
 from gradio.components.base import FormComponent
+from gradio.components.button import Button
 from gradio.events import Events
 from gradio.i18n import I18nData
+from gradio.utils import set_default_buttons
 
 
 @document()
@@ -45,6 +47,7 @@ class DateTime(FormComponent):
         render: bool = True,
         key: int | str | tuple[int | str, ...] | None = None,
         preserved_by_key: list[str] | str | None = "value",
+        buttons: list[Button] | None = None,
     ):
         """
         Parameters:
@@ -63,6 +66,7 @@ class DateTime(FormComponent):
             render: If False, component will not render be rendered in the Blocks context. Should be used if the intention is to assign event listeners now but render the component later.
             key: in a gr.render, Components with the same key across re-renders are treated as the same component, not a new component. Properties set in 'preserved_by_key' are not reset across a re-render.
             preserved_by_key: A list of parameters from this component's constructor. Inside a gr.render() function, if a component is re-rendered with the same key, these (and only these) parameters will be preserved in the UI (if they have been changed by the user or an event listener) instead of re-rendered based on the values provided during constructor.
+            buttons: A list of gr.Button() instances to show in the top right corner of the component. Custom buttons will appear in the toolbar with their configured icon and/or label, and clicking them will trigger any .click() events registered on the button.
         """
         self.type = type
         self.include_time = include_time
@@ -93,6 +97,7 @@ class DateTime(FormComponent):
             value=value,
             interactive=interactive,
         )
+        self.buttons = set_default_buttons(buttons, None)
 
     def preprocess(self, payload: str | None) -> str | float | datetime | None:
         """