+++ gradio/components/dialogue.py
@@ -6,8 +6,10 @@ from typing import Literal
 from gradio_client.documentation import document
 
 from gradio.components.base import Component, server
+from gradio.components.button import Button
 from gradio.data_classes import GradioModel, GradioRootModel
 from gradio.events import Events
+from gradio.utils import set_default_buttons
 
 
 class DialogueLine(GradioModel):
@@ -63,7 +65,7 @@ class Dialogue(Component):
         render: bool = True,
         key: int | str | None = None,
         max_lines: int | None = None,
-        buttons: list[Literal["copy"]] | None = None,
+        buttons: list[Literal["copy"] | Button] | None = None,
         submit_btn: str | bool | None = False,
         ui_mode: Literal["dialogue", "text", "both"] = "both",
     ):
@@ -91,7 +93,7 @@ class Dialogue(Component):
             elem_classes: An optional list of strings that are assigned as the classes of this component in the HTML DOM. Can be used for targeting CSS styles.
             render: If False, component will not render be rendered in the Blocks context. Should be used if the intention is to assign event listeners now but render the component later.
             key: if assigned, will be used to assume identity across a re-render. Components that have the same key across a re-render will have their value preserved.
-            buttons: A list of buttons to show for the component. Currently, the only valid option is "copy". The "copy" button allows the user to copy the text in the textbox. By default, no buttons are shown.
+            buttons: A list of buttons to show for the component. Valid options are "copy" or a gr.Button() instance. The "copy" button allows the user to copy the text in the textbox. Custom gr.Button() instances will appear in the toolbar with their configured icon and/or label, and clicking them will trigger any .click() events registered on the button. By default, no buttons are shown.
             submit_btn: If False, will not show a submit button. If True, will show a submit button with an icon. If a string, will use that string as the submit button text.
             autoscroll: If True, will automatically scroll to the bottom of the textbox when the value changes, unless the user scrolls up. If False, will not scroll to the bottom of the textbox when the value changes.
             ui_mode: Determines the user interface mode of the component. Can be "dialogue" (displays dialogue lines), "text" (displays a single text input), or "both" (displays both dialogue lines and a text input). Defaults to "both".
@@ -125,7 +127,7 @@ class Dialogue(Component):
         self.unformatter = unformatter
         self.separator = separator
         self.color_map = color_map
-        self.buttons = buttons
+        self.buttons = set_default_buttons(buttons, None)
         self.submit_btn = submit_btn
         if not interactive:
             self.info = None