+++ gradio/routes.py
@@ -102,6 +102,7 @@ from gradio.route_utils import (  # noqa: F401
     GradioMultiPartParser,
     GradioUploadFile,
     MultiPartException,
+    NodeProxyCache,
     Request,
     compare_passwords_securely,
     create_lifespan_handler,
@@ -173,7 +174,7 @@ DEFAULT_TEMP_DIR = os.environ.get("GRADIO_TEMP_DIR") or str(
 )
 
 BUILT_IN_THEMES: dict[str, Theme] = {
-    t.name: t
+    t.name: t  # type: ignore
     for t in [
         themes.Base(),
         themes.Default(),
@@ -184,6 +185,7 @@ BUILT_IN_THEMES: dict[str, Theme] = {
         themes.Citrus(),
         themes.Ocean(),
     ]
+    if t.name is not None
 }
 
 
@@ -281,6 +283,7 @@ class App(FastAPI):
     # We're not overriding any defaults here
 
     client = httpx.AsyncClient()
+    proxy_cache = NodeProxyCache(client)
 
     @staticmethod
     async def proxy_to_node(
@@ -312,7 +315,7 @@ class App(FastAPI):
         if mounted_path:
             server_url += mounted_path
 
-        headers = dict(request.headers)
+        headers = {}  # Do not include arbitrary headers from original request so NodeProxyCache can be effective
         headers["x-gradio-server"] = server_url
         headers["x-gradio-port"] = str(python_port)
         headers["x-gradio-mounted-path"] = mounted_path
@@ -321,16 +324,16 @@ class App(FastAPI):
         if os.getenv("GRADIO_LOCAL_DEV_MODE"):
             headers["x-gradio-local-dev-mode"] = "1"
 
-        new_request = App.client.build_request(
-            request.method, httpx.URL(url), headers=headers
-        )
-        node_response = await App.client.send(new_request, stream=True)
+        if (accept_language := request.headers.get("accept-language")) is not None:
+            headers["accept-language"] = accept_language
+
+        proxy_req = App.proxy_cache.ProxyReq(request.method, url, headers)
+        status, response_headers, aiter_raw = await App.proxy_cache.get(proxy_req)
 
         return StreamingResponse(
-            node_response.aiter_raw(),
-            status_code=node_response.status_code,
-            headers=node_response.headers,
-            background=BackgroundTask(node_response.aclose),
+            aiter_raw,
+            status_code=status,
+            headers=response_headers,
         )
 
     def configure_app(self, blocks: gradio.Blocks) -> None:
@@ -444,9 +447,9 @@ class App(FastAPI):
 
         app.configure_app(blocks)
 
-        app.add_middleware(CustomCORSMiddleware, strict_cors=strict_cors)
+        app.add_middleware(CustomCORSMiddleware, strict_cors=strict_cors)  # type: ignore
         app.add_middleware(
-            BrotliMiddleware,
+            BrotliMiddleware,  # type: ignore
             quality=4,
             excluded_handlers=[mcp_subpath],
         )
@@ -867,9 +870,9 @@ class App(FastAPI):
                     }
                 }
 
-                request_properties = path_item["post"]["requestBody"]["content"][
+                request_properties = path_item["post"]["requestBody"]["content"][  # type: ignore
                     "application/json"
-                ]["schema"]["properties"]
+                ]["schema"]["properties"]  # type: ignore
                 for param in endpoint_info.get("parameters", []):
                     param_name = param["parameter_name"]
                     param_type = param.get("type", {})
@@ -882,25 +885,25 @@ class App(FastAPI):
                         param_type = dict(param_type)
                         param_type["type"] = "object"
 
-                    request_properties[param_name] = param_type
+                    request_properties[param_name] = param_type  # type: ignore
 
                     if "example_input" in param:
                         if (
                             "examples"
-                            not in path_item["post"]["requestBody"]["content"][
+                            not in path_item["post"]["requestBody"]["content"][  # type: ignore
                                 "application/json"
                             ]
                         ):
-                            path_item["post"]["requestBody"]["content"][
+                            path_item["post"]["requestBody"]["content"][  # type: ignore
                                 "application/json"
                             ]["examples"] = {"example1": {"value": {}}}
-                        path_item["post"]["requestBody"]["content"]["application/json"][
+                        path_item["post"]["requestBody"]["content"]["application/json"][  # type: ignore
                             "examples"
-                        ]["example1"]["value"][param_name] = param["example_input"]
+                        ]["example1"]["value"][param_name] = param["example_input"]  # type: ignore
 
-                response_properties = path_item["post"]["responses"]["200"]["content"][
+                response_properties = path_item["post"]["responses"]["200"]["content"][  # type: ignore
                     "application/json"
-                ]["schema"]["properties"]
+                ]["schema"]["properties"]  # type: ignore
                 for i, ret in enumerate(endpoint_info.get("returns", [])):
                     ret_name = f"output_{i}" if i > 0 else "output"
                     ret_type = ret.get("type", {})
@@ -913,9 +916,9 @@ class App(FastAPI):
                         ret_type = dict(ret_type)
                         ret_type["type"] = "object"
 
-                    response_properties[ret_name] = ret_type
+                    response_properties[ret_name] = ret_type  # type: ignore
 
-                schema["paths"][f"/run{endpoint_path}"] = path_item
+                schema["paths"][f"/run{endpoint_path}"] = path_item  # type: ignore
 
             return schema