+++ js/atoms/src/index.ts
@@ -11,5 +11,6 @@ export { default as Toolbar } from "./Toolbar.svelte";
 export { default as SelectSource } from "./SelectSource.svelte";
 export { default as IconButtonWrapper } from "./IconButtonWrapper.svelte";
 export { default as FullscreenButton } from "./FullscreenButton.svelte";
+export { default as CustomButton } from "./CustomButton.svelte";
 
 export const BLOCK_KEY = {};