+++ js/highlightedtext/types.ts
@@ -1,4 +1,4 @@
-import type { SelectData } from "@gradio/utils";
+import type { SelectData, CustomButton } from "@gradio/utils";
 import type { LoadingStatus } from "js/statustracker";
 
 export interface HighlightedToken {
@@ -13,10 +13,12 @@ export interface HighlightedTextProps {
 	color_map: Record<string, string>;
 	combine_adjacent: boolean;
 	rtl: boolean;
+	buttons: (string | CustomButton)[] | null;
 }
 
 export interface HighlightedTextEvents {
 	change: never;
 	select: SelectData;
 	clear_status: LoadingStatus;
+	custom_button_click: { id: number };
 }